refactor: move ChooseCheckGroup to page folder
