+++ src/components/CheckStatusInfo/index.ts
@@ -1,0 +1,1 @@
+export { CheckStatusInfo } from './CheckStatusInfo';
