+++ src/page/ChooseCheckGroup/ChooseCheckGroup.test.tsx
