+++ src/page/ChooseCheckGroup/index.ts
